refactor plant card rendering
